Move code about user to user.go
